+++ .github/workflows/misBridgmanApi.yml
@@ -1,0 +1,41 @@
+name: MissBridgman function and API
+
+on:
+  push:
+    branches:
+    - main
+    - work/addApi
+    paths:
+      - api/**
+
+env:
+  AZURE_FUNCTIONAPP_PACKAGE_PATH: '/api/src/MissBridgman' # set this to the path to your web app project, defaults to the repository root
+  DOTNET_VERSION: '7.0.x' # set this to the dotnet version to use
+
+jobs:
+  build-and-deploy:
+    runs-on: ubuntu-latest
+    steps:
+      - name: 'Checkout GitHub Action'
+        uses: actions/checkout@v2
+
+      - name: Setup DotNet ${{ env.DOTNET_VERSION }} Environment
+        uses: actions/setup-dotnet@v1
+        with:
+          dotnet-version: ${{ env.DOTNET_VERSION }}
+
+      - name: 'Resolve Project Dependencies Using Dotnet'
+        shell: bash
+        run: |
+          pushd './${{ env.AZURE_FUNCTIONAPP_PACKAGE_PATH }}'
+          dotnet build --configuration Release --output ./output
+          popd
+
+      - name: 'Run Azure Functions Action'
+        uses: Azure/functions-action@v1
+        id: fa
+        with:
+          app-name: 'missBridgmanApi'
+          slot-name: 'Production'
+          package: '${{ env.AZURE_FUNCTIONAPP_PACKAGE_PATH }}/output'
+          publish-profile: ${{ secrets.AZUREAPPSERVICE_PUBLISHPROFILE_84B34BBBBB0443DAAE4FF41B4804E188 }}
