+++ api/api.sln
@@ -7,9 +7,24 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{B65DABC8-BCB
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{08712E69-7DD4-43D6-9C81-DC0E864A3A90}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MissBridgman", "src\MissBridgman\MissBridgman.csproj", "{3E3C212D-2B43-48A6-9D14-F7475BD84ACD}"
+EndProject
 Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{3E3C212D-2B43-48A6-9D14-F7475BD84ACD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3E3C212D-2B43-48A6-9D14-F7475BD84ACD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3E3C212D-2B43-48A6-9D14-F7475BD84ACD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3E3C212D-2B43-48A6-9D14-F7475BD84ACD}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{3E3C212D-2B43-48A6-9D14-F7475BD84ACD} = {B65DABC8-BCB0-4A12-B383-E63252C188E0}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {EC71D81A-8654-4983-AF44-37C7E8693D41}
